Move table-level functionality to own module
